#20986 Javadoc. Make Expirable implementation inner class. Clean up
