--- dotCMS/src/main/java/com/dotcms/business/Expirable.java
@@ -1,5 +1,0 @@
-package com.dotcms.business;
-
-public interface Expirable {
-    long getTtl();
-}
